modules/aws-cert-lambda/src/requirements.txt
certbot==1.1.0
certbot-dns-route53==1.1.0
pyopenssl==19.1.0
